+++ sdk/python/pulumi_cloudflare/zone.py
@@ -279,17 +279,6 @@ class Zone(pulumi.CustomResource):
                  zone: Optional[pulumi.Input[str]] = None,
                  __props__=None):
         """
-        Provides a Cloudflare Zone resource. Zone is the basic resource for working with Cloudflare and is roughly equivalent to a domain name that the user purchases.
-
-        ## Example Usage
-
-        ```python
-        import pulumi
-        import pulumi_cloudflare as cloudflare
-
-        example = cloudflare.Zone("example", zone="example.com")
-        ```
-
         ## Import
 
         Zone resource can be imported using a zone ID, e.g.
@@ -315,17 +304,6 @@ class Zone(pulumi.CustomResource):
                  args: ZoneArgs,
                  opts: Optional[pulumi.ResourceOptions] = None):
         """
-        Provides a Cloudflare Zone resource. Zone is the basic resource for working with Cloudflare and is roughly equivalent to a domain name that the user purchases.
-
-        ## Example Usage
-
-        ```python
-        import pulumi
-        import pulumi_cloudflare as cloudflare
-
-        example = cloudflare.Zone("example", zone="example.com")
-        ```
-
         ## Import
 
         Zone resource can be imported using a zone ID, e.g.